% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Pvalues_espa.R
\name{Get_Pvalues_espa}
\alias{Get_Pvalues_espa}
\title{P-value calculation using empirical saddlepoint approximation}
\usage{
Get_Pvalues_espa(
  object,
  G,
  N,
  AF,
  cutoff = 2,
  set_nonzero,
  r = 1,
  loco_chr = NULL
)
}
\arguments{
\item{object}{Object returned from Get_Null_Model()}

\item{G}{Genotype matrix (0,1,2)}

\item{N}{Sample size in G}

\item{AF}{Allele frequency of G}

\item{cutoff}{Threshold of sd for utilizing empirical saddlepoint approximation. Default is 2}

\item{set_nonzero}{Index of individuals with non-zero genotypes}

\item{r}{Calibration factor, ratio of GPG and GWG}

\item{loco_chr}{Indicator for leave-one-chromosome-out analysis. Default is NULL}
}
\value{
a vector with three elements: 
 * pval
 * pval_nospa 
 * qstat: variance-adjusted test statistics
 * score: score statistics 
 * var: variance of score
 * converged: indicator for convergence
}
\description{
Test for association between a SNP and a survival object
}
