% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCrossprodGRM}
\alias{getCrossprodGRM}
\title{Get product of GRM and vector}
\usage{
getCrossprodGRM(bVec)
}
\arguments{
\item{bVec}{numeric vector, used for product}
}
\value{
vector
}
\description{
Get product of GRM and vector
}
