% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCrossprodGRM_LOCO}
\alias{getCrossprodGRM_LOCO}
\title{Get product of GRM and vector (LOCO)}
\usage{
getCrossprodGRM_LOCO(bVec)
}
\arguments{
\item{bVec}{numeric vector, used for product}
}
\value{
vector
}
\description{
Get product of GRM and vector (LOCO)
}
