% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCrossprodSig_LOCO}
\alias{getCrossprodSig_LOCO}
\title{Get product of Sigma and vector
Sigma = tau[0] * diag(1/W) + tau[1] * sparsekin + tau[2] * grm}
\usage{
getCrossprodSig_LOCO(bVec, sparsekin, W, tau)
}
\arguments{
\item{bVec}{numeric vector, used for product}

\item{sparsekin}{matrix of kinship coefficients (sparse).}

\item{W}{numeric vector.}

\item{tau}{numeric vector, initial values for variance components.}
}
\value{
vector
}
\description{
Get product of Sigma and vector
Sigma = tau[0] * diag(1/W) + tau[1] * sparsekin + tau[2] * grm
}
