% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCrossprodSig_nok}
\alias{getCrossprodSig_nok}
\title{Get product of Sigma and vector
without sparse kinship matrix
Sigma = tau[0] * diag(1/W) + tau[1] * grm}
\usage{
getCrossprodSig_nok(bVec, W, tau)
}
\arguments{
\item{bVec}{numeric vector, used for product}

\item{W}{numeric vector.}

\item{tau}{numeric vector, initial values for variance components.}
}
\value{
vector
}
\description{
Get product of Sigma and vector
without sparse kinship matrix
Sigma = tau[0] * diag(1/W) + tau[1] * grm
}
