% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDiagOfSigma_LOCO}
\alias{getDiagOfSigma_LOCO}
\title{Get diagonal elements from Sigma (LOCO)
Sigma = tau[0] * diag(1/W) + tau[1] * sparsekin + tau[2] * grm}
\usage{
getDiagOfSigma_LOCO(sparsekin, W, tau)
}
\arguments{
\item{sparsekin}{matrix of kinship coefficients (sparse).}

\item{W}{numeric vector.}

\item{tau}{numeric vector, initial values for variance components.}
}
\value{
A vector with diagonal elements of Sigma
}
\description{
Get diagonal elements from Sigma (LOCO)
Sigma = tau[0] * diag(1/W) + tau[1] * sparsekin + tau[2] * grm
}
