% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDiagOfSigma_nok}
\alias{getDiagOfSigma_nok}
\title{Get diagonal elements from Sigma
without sparse kinship matrix
Sigma = tau[0] * diag(1/W) + tau[2] * grm}
\usage{
getDiagOfSigma_nok(W, tau)
}
\arguments{
\item{W}{numeric vector.}

\item{tau}{numeric vector, initial values for variance components.}
}
\value{
A vector with diagonal elements of Sigma
}
\description{
Get diagonal elements from Sigma
without sparse kinship matrix
Sigma = tau[0] * diag(1/W) + tau[2] * grm
}
