% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMeanDiagOfGRM}
\alias{getMeanDiagOfGRM}
\title{Get mean of diagonal elements from Genetic relationship matrix}
\usage{
getMeanDiagOfGRM()
}
\description{
Get mean of diagonal elements from Genetic relationship matrix
}
