% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOneSNPR}
\alias{getOneSNPR}
\title{Get standardized genotype for a given index}
\usage{
getOneSNPR(i)
}
\arguments{
\item{i}{integer, index of the genotype to be extracted}
}
\value{
A vector of standardized genotypes for Nnomissing individuals
}
\description{
Get standardized genotype for a given index
}
