% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getOneSNPR_nostd}
\alias{getOneSNPR_nostd}
\title{Get unstandardized genotype for a given index}
\usage{
getOneSNPR_nostd(i)
}
\arguments{
\item{i}{integer, index of the genotype to be extracted}
}
\value{
A vector of standardized genotypes for Nnomissing individuals
}
\description{
Get unstandardized genotype for a given index
}
