% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getScore}
\alias{getScore}
\title{Get score value in GLMM}
\usage{
getScore(
  n_nomissing,
  q,
  Y,
  X,
  sparsekin,
  densekin,
  W,
  tau,
  fixtau,
  is_AI,
  tol_pcg,
  maxiter_pcg,
  nrun_trace,
  cutoff_trace
)
}
\arguments{
\item{n_nomissing}{integer, number of non-missing individuals as returned from setgeno()}

\item{q}{integer, total number of sparse and dense correlation matrices}

\item{Y}{column vector, working y.}

\item{X}{data matrix.}

\item{sparsekin}{matrix of kinship coefficients (sparse).}

\item{densekin}{list of dense covariance matrices read from kmatfile.}

\item{W}{numeric vector.}

\item{tau}{numeric vector, initial values for variance components.}

\item{fixtau}{vector with elements 0/1, indicator for whether tau elements are fixed.}

\item{is_AI}{boolean. Whether AI trace estimation will be used.}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm.}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm.}

\item{nrun_trace}{Integer. Number of random vectors used for trace estimation.}

\item{cutoff_trace}{Numeric. Threshold for the coefficient of variation for trace estimation.}
}
\value{
A list with following elements:
* YPAPY
* Trace
* PY
* AI
* cov = XSiX_inv
* tau
* alpha
* dtau
* eta
* XSiX_inv_SiXt
}
\description{
Get score within a variance component estimation iteration
}
