% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getTrace}
\alias{getTrace}
\title{Get trace using Hutchinson’s randomized trace estimator}
\usage{
getTrace(
  n_nomissing,
  q,
  sparsekin,
  densekin,
  Sigma,
  Sigma_iX,
  X,
  W_vec,
  XSiX_inv,
  nrun_trace,
  maxiter_pcg,
  tol_pcg,
  cutoff_trace
)
}
\arguments{
\item{n_nomissing}{integer}

\item{q}{integer, total number of sparse and dense correlation matrices}

\item{sparsekin}{matrix of kinship coefficients (sparse).}

\item{densekin}{list of dense covariance matrices read from kmatfile.}

\item{Sigma}{matrix}

\item{Sigma_iX}{matrix}

\item{X}{matrix}

\item{W_vec}{vector of type arma::fvec.}

\item{XSiX_inv}{matrix, covariance}

\item{nrun_trace}{Integer. Number of random vectors used for trace estimation.}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm.}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm.}

\item{cutoff_trace}{Numeric. Threshold for the coefficient of variation for trace estimation.}
}
\description{
Get trace using Hutchinson’s randomized trace estimator
}
