% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVR.R
\name{getVR}
\alias{getVR}
\title{Calculating variance ratio using genotype subsamples}
\usage{
getVR(
  null_object,
  fit0,
  genfile,
  K,
  outfile,
  n_snp_sample = 30,
  thresh_maf = 0.01,
  verbose = TRUE,
  n_memchunk = 1,
  tol_vr = 0.001,
  tol_pcg = 1e-05,
  maxiter_pcg = 500
)
}
\arguments{
\item{null_object}{Null GLMM model object}

\item{fit0}{Null GLM model object}

\item{genfile}{String. PLINK file to genotype samples}

\item{K}{Matrix (sparse). Kinship or genetic relatedness matrix}

\item{outfile}{String. Output file prefix}

\item{n_snp_sample}{Integer. Number of variants used for variance ratio estimation. Default is 30.}

\item{thresh_maf}{Numeric. MAF threshold. Default is 0.01}

\item{verbose}{Bool. Default is TRUE}

\item{n_memchunk}{Integer. Number of memory chunks. Default is 1.}

\item{tol_vr}{Numeric. Tolerance for variance ratio calculation. Default is 0.001}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm. Default is 1e-5}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm. Default is 500}
}
\value{
Value of variance ratio (with details written to paste0(outfile,"_variance_ratio.txt"))
}
\description{
Calculating variance ratio using genotype subsamples
}
