% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmaiNullModel.R
\name{glmmaiNullModel}
\alias{glmmaiNullModel}
\title{Null Linear Mixed Model using AI-REML}
\usage{
glmmaiNullModel(
  fit0,
  KgenFile = "",
  idstoIncludeFile = "",
  K = NULL,
  KmatFile = "",
  tau = c(1, 0),
  fixtau = c(0, 0),
  maxiter = 500,
  tol = 0.02,
  tol_pcg = 1e-05,
  maxiter_pcg = 500,
  nrun_trace = 30,
  cutoff_trace = 0.0025,
  verbose = FALSE,
  n_memchunk = 1,
  loco = FALSE,
  tol_tau = 1e-05,
  tol_coef = 0.1
)
}
\arguments{
\item{fit0}{GLM object.}

\item{KgenFile}{String. PLINK file to construct the GRM}

\item{idstoIncludeFile}{String. File including a column of rsids to include for GRM construction}

\item{K}{Matrix (sparse). Kinship or genetic relatedness matrix}

\item{KmatFile}{Vector of strings. K matrix file location}

\item{tau}{Vector. Variance component estimates. Default is (1,0)}

\item{fixtau}{Vector. Indicators for fixed value for variance component estimates. Default is (0,0)}

\item{maxiter}{Integer. Max number of iterations to fit the LMM model. Default is 500}

\item{tol}{Numeric. Tolerance for convergence of variance components. Default is 0.02}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm. Default is 1e-5}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm. Default is 500}

\item{nrun_trace}{Integer. Number of random vectors used for trace estimation. Default is 30}

\item{cutoff_trace}{Numeric. Threshold for the coefficient of variation for trace estimation. Default is 0.0025}

\item{verbose}{Indicator for message output. Default is FALSE}

\item{n_memchunk}{Integer. Number of memory chunks for genotype. Default is 1}

\item{loco}{Indicator for leave-one-chromosome-out analysis. Default is FALSE}

\item{tol_tau}{Numeric. Minimum threshold for variance components. Default is 1e-5}

\item{tol_coef}{Numeric. Tolerance for convergence of coefficients. Default is 0.1}
}
\value{
List from getTau() or its variants
}
\description{
Null Linear Mixed Model using AI-REML
}
