% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glmmaiUpdate_plink_LOCO}
\alias{glmmaiUpdate_plink_LOCO}
\title{Update variance components tau in GLMM once}
\usage{
glmmaiUpdate_plink_LOCO(
  n_nomissing,
  q,
  Y,
  X,
  W,
  sparsekin,
  tau,
  fixtau,
  tol,
  tol_pcg,
  maxiter_pcg,
  nrun_trace,
  cutoff_trace
)
}
\arguments{
\item{n_nomissing}{integer, number of non-missing individuals as returned from setgeno()}

\item{q}{integer, total number of sparse and dense correlation matrices}

\item{Y}{column vector, working y.}

\item{X}{data matrix.}

\item{W}{numeric vector.}

\item{sparsekin}{matrix of kinship coefficients (sparse).}

\item{tau}{numeric vector, initial values for variance components.}

\item{fixtau}{vector with elements 0/1, indicator for whether tau elements are fixed.}

\item{tol}{Numeric. Tolerance for estimation of variance components.}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm.}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm.}

\item{nrun_trace}{Integer. Number of random vectors used for trace estimation.}

\item{cutoff_trace}{Numeric. Threshold for the coefficient of variation for trace estimation.}
}
\value{
A List with the following elements:
* tau
* cov = XSiX_inv
* XSiX_inv_SiXt
* alpha
* eta
* Sigma_iX
}
\description{
Variance component update per iteration (LOCO)
}
