% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parallelCrossProdGRM}
\alias{parallelCrossProdGRM}
\title{Get product of GRM and vector parallel worker}
\usage{
parallelCrossProdGRM(bVec)
}
\arguments{
\item{bVec}{numeric vector, used for product}
}
\value{
vector
}
\description{
Get product of GRM and vector parallel worker
}
