% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAPE_Null_Model.R
\name{TAPE_Null_Model}
\alias{TAPE_Null_Model}
\title{Null Model for Test with Adjusted Phenotype using Empirical Saddlepoint Approximation}
\usage{
TAPE_Null_Model(
  formula,
  data = NULL,
  KgenFile = "",
  idstoIncludeFile = "",
  K = NULL,
  KmatFile = "",
  length_out = 9999,
  range = c(-100, 100),
  VRgenFile = NULL,
  idx_ratio = 30,
  tau = c(1, 0),
  fixtau = c(0, 0),
  maxiter = 500,
  tol = 0.02,
  tol_pcg = 1e-05,
  maxiter_pcg = 500,
  nrun_trace = 30,
  cutoff_trace = 0.0025,
  inv_normalize = FALSE,
  verbose = FALSE,
  outFile = "",
  n_memchunk = 1,
  loco = FALSE,
  tol_tau = 1e-05,
  tol_coef = 0.1,
  tol_vr = 0.001,
  thresh_maf_vr = 0.01
)
}
\arguments{
\item{formula}{Formula for the regression model}

\item{data}{Data frame to be used by the formula. Default is NULL.}

\item{KgenFile}{String. PLINK file to construct the GRM}

\item{idstoIncludeFile}{String. File including a column of rsids to include for GRM construction}

\item{K}{Matrix. Sparse matrix for close relatedness}

\item{KmatFile}{String. K matrix file location}

\item{length_out}{Number of samples in resampling. Default is 9999}

\item{range}{Numeric vector with 2 elements. Range for the resampling. Default is c(-100,100)}

\item{VRgenFile}{String. PLINK file of genotypes for estimation of variance ratio. Default is NULL (variance ratio 1)}

\item{idx_ratio}{Integer. Number of variants used for variance ratio estimation. Default is 30.}

\item{tau}{Vector. Variance component estimates. Default is (1,0)}

\item{fixtau}{Vector. Indicators for fixed value for variance component estimates. Default is (0,0)}

\item{maxiter}{Integer. Max number of iterations to fit the LMM model. Default is 500}

\item{tol}{Numeric. Tolerance for convergence of variance components estimation. Default is 0.02}

\item{tol_pcg}{Numeric. Tolerance for the PCG algorithm. Default is 1e-5}

\item{maxiter_pcg}{Integer. Max number of iterations for the PCG algorithm. Default is 500}

\item{nrun_trace}{Integer. Number of random vectors used for trace estimation. Default is 30. If equals NULL, Hutchinson’s randomized trace estimator will not be used.}

\item{cutoff_trace}{Numeric. Threshold for the coefficient of variation for trace estimation. Default is 0.0025}

\item{inv_normalize}{Indicator to perform the inverse normalization for the phentoype. Default is FALSE}

\item{verbose}{Indicator for message output. Default is FALSE}

\item{outFile}{Output file prefix}

\item{n_memchunk}{Integer. Number of memory chunks for genotype. Default is 1}

\item{loco}{Indicator for leave-one-chromosome-out analysis. Default is FALSE}

\item{tol_tau}{Numeric. Minimum threshold for variance components. Default is 1e-5}

\item{tol_coef}{Numeric. Tolerance for convergence of coefficients. Default is 0.1}

\item{tol_vr}{Numeric. Tolerance for variance ratio calculation. Default is 0.001}

\item{thresh_maf_vr}{Numeric. Minimum MAF for markers in variance ratio calculation. Default is 0.01.}
}
\value{
a null model object of class "tape_null", containing the following elements
* resid
* var_resid
* var_ratio
* K_org_emp
* K_1_emp
* K_2_emp
* method
* theta: variance component estimates
* scaled_resid
* X: covariate matrix
* cov: XSigma_inv_X_inv
* W
* XSiX_inv_SiXt
* score0_nok
* mu
* XVX
* converged
* LOCO
* info_LOCO
}
\description{
Null Model for Test with Adjusted Phenotype using Empirical Saddlepoint Approximation
}
