% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAPEtest.R
\name{TAPEtest}
\alias{TAPEtest}
\title{Generating test results from null model}
\usage{
TAPEtest(
  null_object,
  genfile,
  samplefile,
  outfile,
  genfile_format = "bfile",
  thresh_maf = 0,
  thresh_mac = 0,
  snpbatch_size = 20000,
  espa_nt = 9999,
  espa_range = c(-20, 20),
  bgi_file = "1",
  snplist_file = "",
  verbose = FALSE,
  n_memchunk = 1
)
}
\arguments{
\item{null_object}{Null model object}

\item{genfile}{Genotype file prefix}

\item{samplefile}{String, file with a column of IIDs in genfile}

\item{outfile}{Path to output result file}

\item{genfile_format}{(Optional) String, format of genotype file. Default is 'bfile'}

\item{thresh_maf}{(Optional) Threshold of minimum MAF to be analyzed. Default is 0.}

\item{thresh_mac}{(Optional) Threshold of minimum MAC to be analyzed. Default is 0.}

\item{snpbatch_size}{(Optional) Number of SNPs per batch of analysis to be written to outfile. Default is 20000.}

\item{espa_nt}{(Optional) Number of points in empirical saddlepoint approximation, used only if null object is from GMMAT. Default is 9999.}

\item{espa_range}{(Optional) Range of support in approximation, used only if null object is from GMMAT. Default is c(-20,20).}

\item{bgi_file}{(Optional) Path or Suffix type for bgi file. 1="genfile.bgen.bgi", 2="genfile.bgi". Default is 1.}

\item{snplist_file}{(Optional) String, file of a subset of snp ids to be included in test . Default is ''}

\item{verbose}{(Optional) Boolean, indicator for output runtime. Default is FALSE.}

\item{n_memchunk}{(Optional) Integer. Number of memory chunks for PLINK genotype. Default is 1}
}
\value{
Number of tested variants, and test results at outfile
}
\description{
Supports null model objects from TAPE or GMMAT.
}
